Move MultiIcon.java and MultiIconBuilder.java to different package from docking.util (from Docking) to resources (Generic)
