+++ Ghidra/Features/VersionTracking/src/main/java/ghidra/feature/vt/gui/actions/ReplaceFirstMarkupItemAction.java
@@ -22,7 +22,7 @@ import javax.swing.ImageIcon;
 
 import docking.action.MenuData;
 import docking.action.ToolBarData;
-import docking.util.MultiIcon;
+import resources.MultiIcon;
 import ghidra.feature.vt.api.main.VTMarkupItemApplyActionType;
 import ghidra.feature.vt.gui.plugin.VTController;
 import ghidra.feature.vt.gui.plugin.VTPlugin;
