+++ Ghidra/Framework/Generic/src/main/java/resources/MultiIcon.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package docking.util;
+package resources;
 
 import java.awt.*;
 import java.util.ArrayList;
