+++ Ghidra/Framework/Generic/src/main/java/resources/MultiIconBuilder.java
@@ -13,14 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package docking.util;
+package resources;
 
 import java.util.Objects;
 
 import javax.swing.Icon;
 import javax.swing.ImageIcon;
 
-import resources.ResourceManager;
 import resources.icons.TranslateIcon;
 
 /**
